Split PixelTriplets to PixelQuintuplets
